chore: extract ResultsTransformer class and some necessary refactors
