+++ packages/frontend/src/components/DataViz/store/lineChartSlice.ts
@@ -4,8 +4,8 @@ import {
     isLineChartSQLConfig,
 } from '@lightdash/common';
 import { createSlice } from '@reduxjs/toolkit';
-import { SqlRunnerResultsTransformer } from '../../../features/sqlRunner/SqlResultsRunner';
 import { setSavedChartData } from '../../../features/sqlRunner/store/sqlRunnerSlice';
+import { SqlRunnerResultsTransformer } from '../../../features/sqlRunner/transformers/SqlRunnerResultsTransformer';
 import {
     cartesianChartConfigSlice,
     onResults,
