+++ packages/frontend/src/components/DataViz/transformers/ResultsTransformer.ts
@@ -1,9 +1,8 @@
 import {
     DimensionType,
-    MetricType,
+    VizAggregationOptions,
     vizAggregationOptions,
     VizIndexType,
-    type DuckDBSqlFunction,
     type PivotChartData,
     type ResultRow,
     type ResultsRunnerBase,
@@ -15,17 +14,6 @@ import {
     type VizValuesLayoutOptions,
 } from '@lightdash/common';
 import { intersectionBy } from 'lodash';
-import { duckDBFE } from './duckDBQuery';
-
-type GetPivotedResultsArgs = {
-    columns: VizSqlColumn[];
-    rows: RowData[];
-    valuesSql: string[];
-    pivotsSql: string[];
-    groupByColumns: string[];
-    sortsSql: string[];
-    duckDBSqlFunction: DuckDBSqlFunction;
-};
 
 const isResultRows = (rows: (RowData | ResultRow)[]): rows is ResultRow[] => {
     if (rows.length === 0) return false;
@@ -49,65 +37,17 @@ const convertToRowData = (data: ResultRow[]): RowData[] => {
     });
 };
 
-const getPivotedResults = async ({
-    columns,
-    rows,
-    valuesSql,
-    pivotsSql,
-    groupByColumns,
-    sortsSql,
-    duckDBSqlFunction,
-}: GetPivotedResultsArgs) => {
-    const pivotOnSql = pivotsSql.join(', ');
-    const pivotValuesSql = valuesSql.join(', ');
-
-    let query = 'PIVOT results_data';
-    if (pivotsSql.length > 0) {
-        query += ` ON ${pivotOnSql}`;
-    }
-    if (valuesSql.length > 0) {
-        query += ` USING ${pivotValuesSql}`;
-    } else {
-        return {
-            results: [],
-            columns: [],
-        };
-    }
-    if (groupByColumns.length > 0) {
-        query += ` GROUP BY ${groupByColumns.join(', ')}`;
-    }
-    if (sortsSql.length > 0) {
-        query += ` ORDER BY ${sortsSql.join(', ')}`;
-    }
-
-    const pivoted = await duckDBSqlFunction(query, rows, columns);
-
-    const fieldNames = Object.keys(pivoted[0]);
-
-    return {
-        results: pivoted,
-        indexColumns: groupByColumns,
-        valueColumns: fieldNames.filter(
-            (field) => !groupByColumns.includes(field),
-        ),
-    };
-};
-
-export class SqlRunnerResultsTransformer
+export class ResultsTransformer
     implements ResultsRunnerBase<VizSqlCartesianChartLayout>
 {
-    private readonly duckDBSqlFunction: DuckDBSqlFunction;
+    protected readonly rows: RowData[];
 
-    private readonly rows: RowData[];
-
-    private readonly columns: VizSqlColumn[];
+    protected readonly columns: VizSqlColumn[];
 
     constructor(args: {
         rows: (RowData | ResultRow)[];
         columns: VizSqlColumn[];
     }) {
-        this.duckDBSqlFunction = duckDBFE;
-
         this.rows = isResultRows(args.rows)
             ? convertToRowData(args.rows)
             : args.rows;
@@ -165,31 +105,34 @@ export class SqlRunnerResultsTransformer
     }
 
     pivotChartValuesLayoutOptions(): VizValuesLayoutOptions[] {
-        const options: VizValuesLayoutOptions[] = [];
-        for (const column of this.columns) {
+        return this.columns.reduce<VizValuesLayoutOptions[]>((acc, column) => {
             switch (column.type) {
                 case DimensionType.NUMBER:
-                    options.push({
-                        reference: column.reference,
-                        aggregationOptions: vizAggregationOptions,
-                    });
-                    break;
+                    return [
+                        ...acc,
+                        {
+                            reference: column.reference,
+                            aggregationOptions: vizAggregationOptions as any,
+                        },
+                    ];
+
                 case DimensionType.STRING:
                 case DimensionType.BOOLEAN:
-                    options.push({
-                        reference: column.reference,
-                        aggregationOptions: vizAggregationOptions.filter(
-                            (option) =>
-                                option === MetricType.COUNT ||
-                                option === MetricType.COUNT_DISTINCT,
-                        ),
-                    });
-                    break;
+                    return [
+                        ...acc,
+                        {
+                            reference: column.reference,
+                            aggregationOptions: vizAggregationOptions.filter(
+                                (option) =>
+                                    option === VizAggregationOptions.COUNT,
+                            ),
+                        },
+                    ];
+
                 default:
-                    break;
+                    return acc;
             }
-        }
-        return options;
+        }, []);
     }
 
     getPivotChartLayoutOptions(): {
@@ -252,13 +195,13 @@ export class SqlRunnerResultsTransformer
         if (yColumn === undefined) {
             return undefined;
         }
-        const y = [
+        const y: VizSqlCartesianChartLayout['y'] = [
             {
                 reference: yColumn.reference,
                 aggregation:
                     yColumn.type === DimensionType.NUMBER
-                        ? MetricType.SUM
-                        : MetricType.COUNT,
+                        ? VizAggregationOptions.SUM
+                        : VizAggregationOptions.COUNT,
             },
         ];
 
@@ -266,40 +209,6 @@ export class SqlRunnerResultsTransformer
             x,
             y,
             groupBy: undefined,
-        };
-    }
-
-    // args should be rows, columns, values (blocked by db migration)
-    public async getPivotChartData(
-        config: VizSqlCartesianChartLayout,
-    ): Promise<PivotChartData> {
-        const groupByColumns = [config.x.reference];
-        const pivotsSql =
-            config.groupBy === undefined
-                ? []
-                : config.groupBy.map((groupBy) => groupBy.reference);
-        const valuesSql = config.y.map(
-            (y) => `${y.aggregation}(${y.reference})`,
-        );
-        const sortsSql = [`${config.x.reference} ASC`];
-
-        const pivotResults = await getPivotedResults({
-            columns: this.columns,
-            rows: this.rows, // data
-            groupByColumns, // x location
-            valuesSql, // height
-            pivotsSql, // grouping
-            sortsSql,
-            duckDBSqlFunction: this.duckDBSqlFunction,
-        });
-
-        return {
-            results: pivotResults.results,
-            indexColumn: {
-                reference: groupByColumns[0],
-                type: config.x.type,
-            },
-            valuesColumns: pivotResults.valueColumns || [],
         };
     }
 
@@ -321,5 +230,11 @@ export class SqlRunnerResultsTransformer
         }
 
         return mergedLayout;
+    }
+
+    getPivotChartData(
+        _config: VizSqlCartesianChartLayout,
+    ): Promise<PivotChartData> {
+        throw new Error('Method not implemented.');
     }
 }
